+++ TaskExplorer/API/Windows/ProcessHacker/appsup.cpp
@@ -131,13 +131,13 @@ PH_KNOWN_PROCESS_TYPE PhGetProcessKnownTypeEx(quint64 ProcessId, const QString& 
 	return (PH_KNOWN_PROCESS_TYPE)knownProcessType;
 }
 
-BOOLEAN NTAPI PhpSvchostCommandLineCallback(_In_opt_ PPH_COMMAND_LINE_OPTION Option, _In_opt_ PPH_STRING Value, _In_opt_ PVOID Context)
+BOOLEAN NTAPI PhpSvchostCommandLineCallback(_In_opt_ PCPH_COMMAND_LINE_OPTION Option, _In_opt_ PPH_STRING Value, _In_opt_ PVOID Context)
 {
     PPH_KNOWN_PROCESS_COMMAND_LINE knownCommandLine = (PPH_KNOWN_PROCESS_COMMAND_LINE)Context;
 
 	if (Option && Option->Id == 1)
 	{
-		PhSwapReference((PVOID*)&knownCommandLine->ServiceHost.GroupName, Value);
+		PhSwapReference(&knownCommandLine->ServiceHost.GroupName, Value);
 	}
 
     return TRUE;
@@ -797,7 +797,7 @@ VOID PhShellExecuteUserString(
     executeString = PhGetStringSetting(Setting);
 
     // Expand environment strings.
-    PhMoveReference((PVOID*)&executeString, PhExpandEnvironmentStrings(&executeString->sr));
+    PhMoveReference(&executeString, PhExpandEnvironmentStrings(&executeString->sr));
 
     // Make sure the user executable string is absolute. We can't use RtlDetermineDosPathNameType_U
     // here because the string may be a URL.
@@ -814,15 +814,15 @@ VOID PhShellExecuteUserString(
 
             // Make sure the string is absolute and escape the filename.
             if (RtlDetermineDosPathNameType_U(fileName->Buffer) == RtlPathTypeRelative)
-                PhMoveReference((PVOID*)&fileName, PhConcatStrings(4, L"\"", applicationDirectory->Buffer, fileName->Buffer, L"\""));
+                PhMoveReference(&fileName, PhConcatStrings(4, L"\"", applicationDirectory->Buffer, fileName->Buffer, L"\""));
             else
-                PhMoveReference((PVOID*)&fileName, PhConcatStrings(3, L"\"", fileName->Buffer, L"\""));
+                PhMoveReference(&fileName, PhConcatStrings(3, L"\"", fileName->Buffer, L"\""));
 
             // Escape the parameters.
-            PhMoveReference((PVOID*)&fileArgs, PhConcatStrings(3, L"\"", fileArgs->Buffer, L"\""));
+            PhMoveReference(&fileArgs, PhConcatStrings(3, L"\"", fileArgs->Buffer, L"\""));
 
             // Create the escaped execute string.
-            PhMoveReference((PVOID*)&executeString, PhConcatStrings(3, fileName->Buffer, L" ", fileArgs->Buffer));
+            PhMoveReference(&executeString, PhConcatStrings(3, fileName->Buffer, L" ", fileArgs->Buffer));
 
             PhDereferenceObject(fileArgs);
             PhDereferenceObject(fileName);
@@ -831,9 +831,9 @@ VOID PhShellExecuteUserString(
         else
         {
             if (RtlDetermineDosPathNameType_U(executeString->Buffer) == RtlPathTypeRelative)
-                PhMoveReference((PVOID*)&executeString, PhConcatStrings(4, L"\"", applicationDirectory->Buffer, executeString->Buffer, L"\""));
+                PhMoveReference(&executeString, PhConcatStrings(4, L"\"", applicationDirectory->Buffer, executeString->Buffer, L"\""));
             else
-                PhMoveReference((PVOID*)&executeString, PhConcatStrings(3, L"\"", executeString->Buffer, L"\""));
+                PhMoveReference(&executeString, PhConcatStrings(3, L"\"", executeString->Buffer, L"\""));
         }
     }
 
@@ -850,7 +850,7 @@ VOID PhShellExecuteUserString(
         stringTemp = PhCreateString(String);
         stringMiddle = PhGetFileName(stringTemp);
 
-        PhMoveReference((PVOID*)&executeString, PhConcatStringRef3(&stringBefore, &stringMiddle->sr, &stringAfter));
+        PhMoveReference(&executeString, PhConcatStringRef3(&stringBefore, &stringMiddle->sr, &stringAfter));
 
         PhDereferenceObject(stringMiddle);
         PhDereferenceObject(stringTemp);