+++ TaskExplorer/taskexplorer_zh_CN.ts
@@ -3674,6 +3674,10 @@ PartiallyChecked state means apply only to thread std::list.</source>
         <source>Not resolved...</source>
         <translation>未解决...</translation>
     </message>
+    <message>
+        <source>[Unknown SID]</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>CSocketInfo</name>